# backend/artifacts/openapi/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/owner/me": {
      "get": {
        "operationId": "profile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Идентификатор, телефон и число питомцев."
          },
          "401": {
            "description": "Bearer token отсутствует или невалиден."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Профиль текущего владельца",
        "tags": [
          "Owner profile"
        ]
      }
    },
...
    "/v1/owner/pets": {
      "get": {
        "operationId": "OwnerPetMvp_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OwnerPetMvpDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Owner pets"
        ]
      },
      "post": {
        "operationId": "OwnerPetMvp_create",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "required": [
                  "name",
                  "species"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 120
                  },
                  "species": {
                    "type": "string",
                    "enum": [
                      "DOG",
                      "CAT",
                      "OTHER"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OwnerPetMvpDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Owner pets"
        ]
      }
    },
    "/v1/owner/pets/{petId}": {
      "get": {
        "operationId": "OwnerPetMvp_read",
        "parameters": [
          {
            "name": "petId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OwnerPetMvpDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Owner pets"
        ]
      }
    },
    "/v1/auth/session": {
      "get": {
        "operationId": "EffectiveSession_read",
        "parameters": [],
        "responses": {
          "200": {
            "description": "UX hint only; authorization is always evaluated server-side.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EffectiveOwnerSessionDto"
                }
              }
            }
          },
          "401": {
            "description": "Opaque session or legacy JWT is invalid, expired or revoked.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Authoritative effective session read for opaque Owner session or legacy staff JWT",
        "tags": [
          "Authentication"
        ]
      }
    },
...
  },
  "info": {
    "title": "VetHelp Booking Core API",
    "description": "Контракт локального удержания слотов, подтверждения записи клиникой и освобождения hold.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "scheme": "bearer",
        "bearerFormat": "opaque Owner session or legacy staff JWT",
        "type": "http",
        "description": "Opaque vh_ Owner session or compatibility JWT for staff/system roles. Authorization and expiry are evaluated server-side for Owner sessions."
      }
    },
    "schemas": {
...
      "AuthErrorDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "enum": [
              "VALIDATION_ERROR",
              "OTP_INVALID",
              "OTP_EXPIRED",
              "OTP_RESEND_COOLDOWN",
              "OTP_RATE_LIMITED",
              "OTP_TEMPORARILY_BLOCKED",
              "OTP_CHALLENGE_NOT_FOUND",
              "OTP_ALREADY_USED",
              "OTP_PROVIDER_UNAVAILABLE",
              "OTP_PROVIDER_TIMEOUT",
              "OTP_PROVIDER_OUTCOME_UNKNOWN",
              "INVALID_SESSION",
              "INTERNAL_ERROR"
            ]
          },
          "message": {
            "type": "string"
          },
          "correlationId": {
            "type": "string",
            "format": "uuid"
          },
          "retryAt": {
            "type": "string",
            "format": "date-time"
          },
          "attemptsRemaining": {
            "type": "number",
            "minimum": 0,
            "maximum": 5
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
...
      "OwnerPetMvpDto": {
        "type": "object",
        "properties": {
          "petId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 120
          },
          "species": {
            "type": "string",
            "enum": [
              "DOG",
              "CAT",
              "OTHER"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "petId",
          "name",
          "species",
          "createdAt",
          "updatedAt"
        ]
      },
      "ApiErrorDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 403
          },
          "code": {
            "type": "string",
            "example": "PET_OWNERSHIP_MISMATCH"
          },
          "message": {
            "type": "string",
            "example": "Pet ownership mismatch"
          },
          "correlationId": {
            "type": "string",
            "format": "uuid"
          },
          "details": {
            "type": "object"
          }
        },
        "required": [
          "statusCode",
          "code",
          "message"
        ]
      },
      "EffectiveSessionScopeDto": {
        "type": "object",
        "properties": {
          "clinicId": {
            "type": "string",
            "format": "uuid"
          },
          "locationId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "clinicId",
          "locationId"
        ]
      },
      "EffectiveOwnerSessionDto": {
        "type": "object",
        "properties": {
          "subjectId": {
            "type": "string",
            "format": "uuid"
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "effectiveCapabilities": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "clinicScopes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EffectiveSessionScopeDto"
            }
          }
        },
        "required": [
          "subjectId",
          "roles",
          "effectiveCapabilities",
          "clinicScopes"
        ]
      },
...
    }
  }
}